refactor: implement network toast
