+++ package.json
@@ -59,7 +59,7 @@
     "fathom-client": "3.1.0",
     "formik": "2.2.9",
     "jotai": "1.2.2",
-    "jotai-query-toolkit": "0.1.6",
+    "jotai-query-toolkit": "0.1.8",
     "just-debounce-it": "1.5.0",
     "lru-cache": "6.0.0",
     "mdi-react": "7.5.0",
