+++ src/pages/block/[hash].tsx
@@ -26,6 +26,8 @@ import {
   useBlockTxsCurrentlyInView,
 } from '../../hooks/currently-in-view-hooks';
 import { PageWrapper } from '@components/page-wrapper';
+import { useNetworkToast } from '@common/hooks/use-network-toast';
+import { NetworkModeToast } from '@components/network-mode-toast';
 
 interface BlockSinglePageData {
   hash: string;
@@ -35,11 +37,13 @@ interface BlockSinglePageData {
 const BlockSinglePage: NextPage<BlockSinglePageData> = ({ error, hash }) => {
   const block = useBlockCurrentlyInView();
   const transactions = useBlockTxsCurrentlyInView();
+  useNetworkToast(error || !block || !transactions);
   if (error || !block || !transactions) {
     return (
       <>
         <Meta title="Block hash not found" />
-        <BlockNotFound isPending={validateTxId(hash)} />;
+        <BlockNotFound isPending={validateTxId(hash)} />
+        <NetworkModeToast />
       </>
     );
   }
@@ -101,6 +105,7 @@ const BlockSinglePage: NextPage<BlockSinglePageData> = ({ error, hash }) => {
       {transactionsWithoutCoinbase?.length ? (
         <TransactionList mt="extra-loose" transactions={transactionsWithoutCoinbase} />
       ) : null}
+      <NetworkModeToast />
     </PageWrapper>
   );
 };
@@ -113,6 +118,7 @@ BlockSinglePage.getInitialProps = ctx => {
     error: false,
   };
 };
+
 export default withInitialQueries<Block, BlockSinglePageData>(BlockSinglePage, pageAtomBuilders)(
   getBlockPageQueries,
   getBlockPageQueryProps
