+++ src/pages/transactions.tsx
@@ -1,19 +1,19 @@
 import React from 'react';
+import type { NextPage } from 'next';
+import { withInitialQueries } from 'jotai-query-toolkit/nextjs';
 import { Box } from '@stacks/ui';
-
 import { Title } from '@components/typography';
 import { Meta } from '@components/meta-head';
 import { TabbedTransactionList } from '@components/tabbed-transaction-list';
-
-import { withInitialQueries } from 'jotai-query-toolkit/nextjs';
 import { PageWrapper } from '@components/page-wrapper';
 import { DEFAULT_LIST_LIMIT } from '@common/constants';
 import { getTransactionsPageQueries } from '@common/page-queries/transactions';
 import { pageAtomBuilders } from '@common/page-queries/extra-initial-values';
-
-import type { NextPage } from 'next';
+import { useNetworkToast } from '@common/hooks/use-network-toast';
+import { NetworkModeToast } from '@components/network-mode-toast';
 
 const TransactionsPage: NextPage = () => {
+  useNetworkToast();
   return (
     <PageWrapper>
       <Meta title="Recent transactions" />
@@ -23,6 +23,7 @@ const TransactionsPage: NextPage = () => {
         </Title>
         <TabbedTransactionList infinite limit={DEFAULT_LIST_LIMIT} />
       </Box>
+      <NetworkModeToast />
     </PageWrapper>
   );
 };
